requirements.txt
telethon>=1.28.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
